rename to "from_attributes", fix tests
