+++ pydantic_core/_types.py
@@ -29,6 +29,7 @@ class ConfigSchema(TypedDict, total=False):
     extra: Literal['allow', 'forbid', 'ignore']
     model_full: bool  # default: True
     allow_population_by_field_name: bool
+    from_attributes: bool
 
 
 class DictSchema(TypedDict, total=False):
